Se agrega callback

+++ src/app/components/final/final.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { SessionService } from 'src/app/services/session/session.service';
 import { Rutas } from 'src/app/model/RutasUtil';
+import { sesionModel } from 'src/app/model/sesion/SessionPojo';
 
 @Component({
   selector: 'app-final',
@@ -10,10 +11,23 @@ import { Rutas } from 'src/app/model/RutasUtil';
 })
 export class FinalComponent implements OnInit {
 
-  constructor(private router: Router, private sesion: SessionService) { }
+  constructor(private router: Router, private sesion: SessionService, private session: SessionService) { }
+
+  object: sesionModel;
+  routeToReturn: string;
 
   ngOnInit() {
-    sessionStorage.clear();
+    this.getDataToRedirect();
   }
 
+  getDataToRedirect() {
+    this.object = this.session.getObjectSession();
+    if ( this.object !== undefined && this.object !== null && this.object.callback) {
+      const returnUrl = this.object.callback;
+      sessionStorage.clear();
+      setTimeout( () => {
+        window.location.href = returnUrl;
+      }, 5000 );
+    }
+  }
 }
